requirements.txt

langchain
langgraph
langchain-community
langchain-openai
pydantic

chromadb



transformers
torch



pillow
openpyxl
pypdf
pdf2image



requests
python-dotenv
numpy


matplotlib
seaborn
scikit-learn

requests>=2.31.0
